pyproject.toml
[project]
name = "hyperfunc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "trulite", email = "satya@bluecopa.com" }
]
requires-python = ">=3.10"
dependencies = [
    "peewee>=3.17.0",
    "pydantic>=2.12.4",
    "torch>=2.9.1",
    "torcheggroll>=0.1.0",
]

[project.optional-dependencies]
llm = ["litellm>=1.0"]
postgres = ["psycopg2-binary>=2.9"]
otlp = [
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]
otlp-http = [
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=0.24.0",
    "pytest-dotenv>=0.5.2",
    "jax>=0.4.0",
    "litellm>=1.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "testcontainers>=4.0.0",
    "requests>=2.28.0",
]

[tool.pytest.ini_options]
env_files = [".env.local"]
markers = [
    "integration: marks tests as integration tests requiring external services (deselect with '-m \"not integration\"')",
]
